refactor: create product 리팩터링
