+++ src/main/java/kitchenpos/products/application/ProductService.java
@@ -33,20 +33,8 @@ public class ProductService {
     }
 
     @Transactional
-    public Product create(final Product request) {
-        final BigDecimal price = request.getPrice();
-        if (Objects.isNull(price) || price.compareTo(BigDecimal.ZERO) < 0) {
-            throw new IllegalArgumentException();
-        }
-        final String name = request.getName();
-        if (Objects.isNull(name) || purgomalumClient.containsProfanity(name)) {
-            throw new IllegalArgumentException();
-        }
-        final Product product = new Product();
-        product.setId(UUID.randomUUID());
-        product.setName(name);
-        product.setPrice(price);
-        return productRepository.save(product);
+    public Product create(final CreateProductRequest request) {
+        return productRepository.save(new Product(UUID.randomUUID(), request.getDisplayedName(), request.getPrice()));
     }
 
     @Transactional
