+++ src/main/java/kitchenpos/products/domain/Price.java
@@ -1,10 +1,16 @@
-package kitchenpos.products.tobe.domain;
+package kitchenpos.products.domain;
 
+import javax.persistence.Embeddable;
 import java.math.BigDecimal;
 import java.util.Objects;
 
+@Embeddable
 public class Price {
-    private final BigDecimal price;
+    private BigDecimal price;
+
+    protected Price() {
+    }
+
     public Price(long price) {
         if (price < 0) {
             throw new IllegalArgumentException(String.format("가격은 0원 이상이어야 합니다. 현재 값: %s", price));
